Refactored expression for generating ascii printable char bytes into a helper func.
Refactored plaintext_scoring function to accept a reference to an Option<AsciiString> to account for unencodable plaintexts.

Remove use of unwrapping the result of from_ascii in favour of using the Result::ok function to convert a Result into an Option to handle candidate plaintexts that aren't valid ASCII strings in challenge 3.